api/requirements.txt
requests==2.25.0
python-dotenv
tqdm==4.54.0
pandas==1.1.5
flask
flask_cors
wget
opencv-python
Pillow